Create Big Batch of symbols
The sans font is essentially done. Could use a lot of little tweaks though.
Alternate svg files, an html file for testing the characters, and a build script, based on what I did with openmoji.

The Quivira-based serif font is far from done. (and the build script needs tweaking too.)